overrides base tile pruning call to fix cache conflicts
The tile pruning function provided by Leaflet retains parent and child
tiles, make them immediately available after zoom actions.  This
conflicts with our caching strategy and results in the parent and child
tiles being left in the map.  This is addressed by overriding the
baseline _pruneTiles function with a version that doesn't retain
parents and children.